// src/app/models/servicio9.ts
export interface IdFichaClinica {
  idFichaClinica: number;
  fechaHora: string;
  motivoConsulta: string;
  diagnostico: string;
  observacion: string;
  idLocal: IdLocal;
  idEmpleado: IdEmpleado;
  idCliente: IdCliente;
  idTipoProducto: IdTipoProducto;
  fechaHoraCadena: string;
  fechaHoraCadenaFormateada: string;
  fechaDesdeCadena?: any;
  fechaHastaCadena?: any;
  todosLosCampos?: any;
}

export interface IdEmpleado2 {
  idPersona: number;
  nombre: string;
  apellido: string;
  email: string;
  telefono: string;
  seguroMedico?: any;
  seguroMedicoNumero?: any;
  ruc: string;
  cedula: string;
  tipoPersona: string;
  usuarioLogin?: any;
  idLocalDefecto?: any;
  flagVendedor: string;
  flagTaxfree?: any;
  flagExcepcionChequeoVenta?: any;
  observacion?: any;
  direccion?: any;
  idCiudad?: any;
  tipoCliente: string;
  fechaHoraAprobContrato?: any;
  soloUsuariosDelSistema?: any;
  soloPersonasTaxfree?: any;
  nombreCompleto: string;
  limiteCredito: number;
  fechaNacimiento: string;
  soloProximosCumpleanhos?: any;
  todosLosCampos?: any;
  incluirLimiteDeCredito?: any;
  deuda?: any;
  saldo?: any;
  creditos?: any;
}

export interface IdEmpleadoSecundario {
  idPersona: number;
  nombre: string;
  apellido: string;
  email: string;
  telefono: string;
  seguroMedico?: any;
  seguroMedicoNumero?: any;
  ruc: string;
  cedula: string;
  tipoPersona: string;
  usuarioLogin?: any;
  idLocalDefecto?: any;
  flagVendedor: string;
  flagTaxfree?: any;
  flagExcepcionChequeoVenta?: any;
  observacion?: any;
  direccion?: any;
  idCiudad?: any;
  tipoCliente: string;
  fechaHoraAprobContrato?: any;
  soloUsuariosDelSistema?: any;
  soloPersonasTaxfree?: any;
  nombreCompleto: string;
  limiteCredito: number;
  fechaNacimiento: string;
  soloProximosCumpleanhos?: any;
  todosLosCampos?: any;
  incluirLimiteDeCredito?: any;
  deuda?: any;
  saldo?: any;
  creditos?: any;
}
// ...
export interface Usuario {
  idPersona: number;
  nombre: string;
  apellido: string;
  email?: any;
  telefono?: any;
  seguroMedico?: any;
  seguroMedicoNumero?: any;
  ruc?: any;
  cedula?: any;
  tipoPersona: string;
  usuarioLogin: string;
  idLocalDefecto: IdLocalDefecto;
  flagVendedor: string;
  flagTaxfree?: any;
  flagExcepcionChequeoVenta?: any;
  observacion?: any;
  direccion?: any;
  idCiudad?: any;
  tipoCliente: string;
  fechaHoraAprobContrato?: any;
  soloUsuariosDelSistema?: any;
  soloPersonasTaxfree?: any;
  nombreCompleto: string;
  limiteCredito: number;
  fechaNacimiento?: any;
  soloProximosCumpleanhos?: any;
  todosLosCampos?: any;
  incluirLimiteDeCredito?: any;
  deuda?: any;
  saldo?: any;
  creditos?: any;
}

export interface Servicio9 {
  idServicio: number;
  flagEstado: string;
  fechaHora: string;
  observacion: string;
  presupuesto: number;
  idFichaClinica: IdFichaClinica;
  idEmpleado: IdEmpleado2;
  idEmpleadoSecundario: IdEmpleadoSecundario;
  usuario: Usuario;
  fechaHoraCadena: string;
  fechaDesdeCadena?: any;
  fechaHastaCadena?: any;
  estado: string;
}

export interface ListadoServicio9 {
  lista: Servicio9[];
  totalDatos: number;
}